src/requirements.txt
Flask
Flask_wtf
psycopg2-binary
pytest
colorama
coverage
pytest coverage
beautifulsoup4